🔧Edited index.js 🚚 moved to routes 🔧update refs
